backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
psycopg2-binary==2.9.3
sentry-sdk==1.16.0
Pillow==10.0.1
django-filter==23.3
